CAUTION: BREAKING; Altered mapping and stored data.
In working on a wrapper for WP_Query, I recognized that SP should store `post_password` and it should also store author data differently. The updates to post_author are breaking, and you should flush and reindex your data after applying this commit. We'll automate that process prior to releasing this version.